SEQNG-335 Spilt components files
